Created blif_ready.py
This should be the beginning steps of parsing through a .blif file and writing out a corresponding Singular file to then execute. I still need to get the RTTO on the variables and that has not really been worked on yet. This is designed specifically for a multiplier .blif as generated by abc on the CADE labs. I was testing the concept on a 4-bit multiplier and it seems to have some of the kinks worked out and will need some more work still.