Uploading blif_parser.py file and example files
The blif_parser.py can take in the output of a "gen" command in abc after it has been read again by abc and the "sweep" command performed prior to writing it out again. This Multi4.blif file was created according to those steps in abc and the write.sing file is the output from giving it as an argument to the blif_parser.py.